gwselect-package {lagr}	R Documentation

Variable selection for a geographically weighted regression model, using the LASSO

Description

Variable selection for a geographically weighted regression model, using the LASSO

Details

Package: 	 gwselect

Type: 	 Package

Version: 	 0.1

Date: 	 2012-08-24

License: 	 GPL-3

Author(s)

Wesley Brooks Maintainer: Wesley Brooks <wesley@somesquares.com>

References

~~ Literature or other references for background information ~~

See Also

~~ Optional links to other man pages, e.g. ~~ ~~ <pkg> ~~

[Package lagr version 0.1 Index]